requirements.txt
# Agentic Payment Operations - Python dependencies
# LLM: set GEMINI_API_KEY to use Gemini 2.5 Flash; system falls back to heuristics without it.
google-generativeai>=0.8.0
streamlit>=1.28.0
pandas>=2.0.0
